+++ src/components/NewInvestment.js
@@ -1,5 +1,5 @@
 import { Button, useDisclosure, Spacer } from '@nextui-org/react';
-import InvestmentModal from './InvestmentModal';
+import NewInvestmentModalContent from './Modal/NewInvestmentModalContent';
 import { useTranslation } from 'react-i18next';
 import { NewAccountTypeModalContent } from './Modal/NewAccountTypeModalContent';
 import EmptyModal from './Modal/EmptyModal';
@@ -16,7 +16,10 @@ export default function NewInvestment({ isInvestment }) {
       <Spacer y={4} />
       <Button onPress={onOpen} color='primary'>{t('newInvestment.add')}</Button>
       {
-        isInvestment ? <InvestmentModal isOpen={isOpen} onOpenChange={onOpenChange} isEdit={false} /> :
+        isInvestment ?
+          <EmptyModal isOpen={isOpen} onOpenChange={onOpenChange} hideCloseButton={false}>
+            <NewInvestmentModalContent isEdit={false} />
+          </EmptyModal> :
           <EmptyModal isOpen={isOpen} onOpenChange={onOpenChange} >
             <NewAccountTypeModalContent isEdit={false} />
           </EmptyModal>
